Revert 886618, this actually causes more crashes on 64 bit systems. Thanks to Matthias for testing the revert. BUG: 175663 BUG: 175667
svn path=/trunk/KDE/kdepim/; revision=886941

+++ kmmsgdict.cpp
@@ -452,7 +452,7 @@ int KMMsgDict::readFolderIds( FolderStorage& storage )
   KMMsgDictREntry *rentry = new KMMsgDictREntry(count);
 
   for (unsigned int index = 0; index < count; index++) {
-    unsigned long msn;
+    quint32 msn;
 
     bool readOk = fread(&msn, sizeof(msn), 1, fp);
     if (swapByteOrder)
